+++ localization/autoware_ndt_scan_matcher/include/autoware/ndt_scan_matcher/ndt_omp/ndt_struct.hpp
@@ -91,7 +91,7 @@ struct NdtParams
 {
   double trans_epsilon{};
   double step_size{};
-  double resolution{};
+  float resolution{};
   int max_iterations{};
   NeighborSearchMethod search_method{};
   int num_threads{};
